+++ src/CombatSearch_BestResponse.cpp
@@ -31,7 +31,7 @@ void CombatSearch_BestResponse::recurse(const GameState & state, int depth)
     ActionSetAbilities legalActions;
     generateLegalActions(state, legalActions, m_params);
     
-    for (size_t a(0); a < legalActions.size(); ++a)
+    for (int a(0); a < legalActions.size(); ++a)
     {
         int ri = (int)legalActions.size() - 1 - a;
